fix: use `size_t`

+++ src/messages.cpp
@@ -529,7 +529,7 @@ void Messages::dialog::init( ui_adaptor &ui )
     w_fh_width = w_width;
     w_fh_x = w_x;
     help_text = filter_help_text( w_fh_width - border_width * 2 );
-    w_fh_height = help_text.size() + static_cast<unsigned long>( border_width ) * 2;
+    w_fh_height = static_cast<int>( help_text.size() ) + border_width * 2;
     w_fh_y = w_y + w_height - w_fh_height;
     w_filter_help = catacurses::newwin( w_fh_height, w_fh_width, point( w_fh_x, w_fh_y ) );
